Adding my work
I added a test folder and added the tests that @jredisme had already added to it. I also turned the functions of the program I am over into their own separate functions. One thing that I think would help is if we turned the running of the machine into its own function that the main called, so that we can write more tests against that. Also @JarenMease note that I wrote a test case to test the read_program, and it is failing. I will leave it to you to fix.